requirements.txt
pillow==4.2.1
numpy==1.14.2
chainer==4.5.0
chainercv==0.10.0